+++ packages/utils/.gitignore
@@ -1,6 +1,1 @@
-/*.js
-/**/*.js
-!*.test.js
-!/**/*.test.js
-!/**/__testfixtures__/*.js
-!/**/__snapshots__/*.js
+lib/
