+++ packages/utils/__tests__/is-local-path.test.ts
@@ -1,7 +1,7 @@
 "use strict";
 
 import * as path from "path";
-import { isLocalPath } from "../path-utils";
+import { isLocalPath } from "../src/path-utils";
 
 describe("is-local-path", () => {
 	it("returns true for paths beginning in the current directory", () => {
