+++ packages/utils/__tests__/npm-packages-exists.test.ts
@@ -1,5 +1,5 @@
-import npmPackagesExists from "../npm-packages-exists";
-import { resolvePackages } from "../resolve-packages";
+import npmPackagesExists from "../src/npm-packages-exists";
+import { resolvePackages } from "../src/resolve-packages";
 
 jest.mock("../npm-exists");
 jest.mock("../resolve-packages");
