+++ packages/utils/tsconfig.json
@@ -1,3 +1,7 @@
 {
-  "extends": "../../tsconfig.packages.json"
+  "extends": "../../tsconfig.packages.json",
+  "compilerOptions": {
+    "outDir": "./lib"
+  },
+  "include": ["./src"]
 }
